refactor Server code
