+++ Server/src/model/KeyLogger.java
@@ -1,4 +1,4 @@
-package src;
+package model;
 
 import com.github.kwhat.jnativehook.keyboard.NativeKeyEvent;
 import com.github.kwhat.jnativehook.keyboard.NativeKeyListener;
